requirements.txt
flask
pandas
numpy
scikit-learn
requests
nltk
faiss-cpu
python-dotenv
rank-bm25
scikit-learn==1.5.0


flask-cors


transformers==4.40.0  
sentence-transformers==2.7.0  
torch==2.3.0